requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
codegen==1.0
cryptography==2.1.4
defusedxml==0.5.0
distro==1.3.0
Django==2.1.3
django-allauth==0.38.0
django-email-users==1.1.0
django-extensions==2.1.4
django-rest-auth==0.9.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
dnspython==1.15.0
eventlet==0.24.1
greenlet==0.4.15
gunicorn==19.9.0
httplib2==0.9.2
idna==2.6
include==0.2.2
jeepney==0.4
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
monotonic==1.5
oauth==1.0.1
oauthlib==2.1.0
olefile==0.45.1
patterns==0.3
pbr==5.1.1
pexpect==4.2.1
Pillow==5.1.0
pkg-resources==0.0.0
protobuf==3.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.6.4
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-debian==0.1.32
python-decouple==3.1
python-openid==2.2.5
python3-openid==3.1.0
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
testresources==2.0.1
urllib3==1.22
virtualenv==16.0.0
wadllib==1.3.2
zope.interface==4.3.2
